Extract plot from RuntimeProvider.finalize(), reap pipeline, and spawn env injection
Deletion-pass step 7 of pl-3a79 (warren-39e3). Pulls the plot merge/commit
out of the finalize seam and the plot dispatch env out of spawn, across both
runtime providers and the shared contract in one commit so typecheck stays green.

- contract.ts / finalize-deltas.ts: drop PlotDelta, the 'plot' mirror/commit
  union members, plot_merge/plot_commit FinalizeStage, and the PLOT_ID/PLOT_ACTOR
  env doc; FinalizeIntent.commit narrows to 'seeds'[].
- runtime/local/finalize.ts + runtime/k8s/finalize*: drop plot merge/commit
  collection and wire validation.
- reap pipeline/run/types/util: drop plot state fields, plot_merge/plot_commit
  ReapStep, and the plot terminal-event assembly; delete reap/plot-merge.ts and
  stagePlotForCommit in stage.ts.
- spawn dispatch/types/index: drop the hasPlot guard, plotId input,
  emitRunDispatchedToPlot, and PLOT_ID/PLOT_ACTOR env; delete spawn/plot-append.ts.
- relocate resolveDispatcherHandle/DEFAULT_DISPATCHER_HANDLE to
  src/plots/dispatcher-handle.ts (plot HTTP handlers still consume them until
  plot-core deletion, warren-b8bf).

check:all green (12/12).
