+++ .mulch/expertise/runtime.jsonl
@@ -1,2 +1,3 @@
 {"type":"failure","classification":"tactical","recorded_at":"2026-07-14T16:00:08.832Z","evidence":{"seeds":"warren-1ca1","commit":"edadb11e251786c78cde65e39f8fb8f15a1b693d"},"description":"Every spawnRun wrapper must thread the boot-resolved runtimeProvider: bootScheduler, createPlanRunSpawn, and createMergePollerDispatch omitted it, so under WARREN_RUNTIME=k8s all cron/scheduled/plan-run/planner dispatches fell back to LocalProvider and failed never_started against the absent burrow socket, retrying every 60s tick (~200 failed runs in 3h on GKE). Same class as warren-c531 (bridges/poller/watchdog).","resolution":"Threaded runtimeProvider through all three spawn wrappers + boot wiring (server/main/index.ts, detector-wiring.ts); regression tests assert the provider instance reaches spawnRun. When adding a new spawnRun call site, mirror how POST /runs threads runtimeProvider.","id":"mx-f0322b"}
 {"type":"failure","classification":"tactical","recorded_at":"2026-07-14T16:00:09.002Z","evidence":{"seeds":"warren-035c","commit":"edadb11e251786c78cde65e39f8fb8f15a1b693d"},"description":"warrenCommitIdentityArgs() ('-c user.name=...') loses to inherited GIT_AUTHOR_NAME/GIT_AUTHOR_EMAIL env — git exports those to pre-commit hooks, so this repo's own hook (check:all -> bun test) deterministically fails clone-apply.test.ts 'commits as the warren bot' on every local commit. Repro: GIT_AUTHOR_NAME=x bun test src/runs/reap/clone-apply.test.ts.","resolution":"Open as warren-035c: add warrenCommitIdentityEnv() and pass explicit GIT_AUTHOR_/GIT_COMMITTER_ env at the four bot-commit sites. Until it lands, a commit whose direct check:all run is green may need --no-verify with justification.","id":"mx-27c5be"}
+{"type":"decision","classification":"tactical","recorded_at":"2026-07-21T19:04:46.522Z","evidence":{"commit":"2bd7bbb5da6d892e354289defcf341797a86199a"},"dir_anchors":["src/runs/reap","src/runs/spawn","src/runtime/k8s","src/server/handlers/plots"],"title":"Plot extracted from RuntimeProvider.finalize + reap + spawn (warren-39e3)","rationale":"Deletion-pass step 7 of pl-3a79: plot must leave the finalize seam before plot core is deleted (warren-b8bf). Both providers + shared contract move in one commit or typecheck breaks.","id":"mx-d3ec70"}
