+++ src/runs/reap/run.ts
@@ -204,12 +204,6 @@ export async function reapRun(input: ReapRunInput): Promise<ReapRunResult> {
 			seedIdClosed: state.seedIdClosed,
 			committed: state.seedsCommitted,
 		},
-		plot: {
-			eventsAppended: state.plotEventsAppended,
-			plotsUpdated: state.plotsUpdated,
-			mirrored: state.plotEventsMirrored,
-			committed: state.plotCommitted,
-		},
 		branchPushed: state.branchPushed,
 		commitsAhead: state.commitsAhead,
 		// warren-89b0: distinguish a deliberate no-op (succeeded, non-alarming)
@@ -267,10 +261,6 @@ export async function reapRun(input: ReapRunInput): Promise<ReapRunResult> {
 			seedsCreated: state.seedsCreated,
 			seedIdClosed: state.seedIdClosed,
 			seedsCommitted: state.seedsCommitted,
-			plotEventsAppended: state.plotEventsAppended,
-			plotsUpdated: state.plotsUpdated,
-			plotEventsMirrored: state.plotEventsMirrored,
-			plotCommitted: state.plotCommitted,
 			branchPushed: state.branchPushed,
 			commitsAhead: state.commitsAhead,
 			prUrl: state.prUrl,
@@ -295,10 +285,6 @@ export async function reapRun(input: ReapRunInput): Promise<ReapRunResult> {
 		seedsClosed: state.seedsClosed,
 		seedsCreated: state.seedsCreated,
 		seedIdClosed: state.seedIdClosed,
-		plotEventsAppended: state.plotEventsAppended,
-		plotsUpdated: state.plotsUpdated,
-		plotEventsMirrored: state.plotEventsMirrored,
-		plotCommitted: state.plotCommitted,
 		seedsCommitted: state.seedsCommitted,
 		branchPushed: state.branchPushed,
 		commitsAhead: state.commitsAhead,
