+++ src/runs/reap/stage.test.ts
@@ -14,7 +14,7 @@ import {
 	setup,
 } from "./test-helpers.ts";
 
-describe("reapRun commit-through-reap sub-steps (warren-343a + warren-7ecc)", () => {
+describe("reapRun commit-through-reap sub-steps (warren-7ecc)", () => {
 	let ctx: Ctx;
 
 	beforeEach(async () => {
@@ -23,253 +23,6 @@ describe("reapRun commit-through-reap sub-steps (warren-343a + warren-7ecc)", ()
 
 	afterEach(async () => {
 		await ctx.db.close();
-	});
-
-	/* ------------------------------------------------------------------ */
-	/* warren-343a: commit-through-reap for .plot/                         */
-	/* ------------------------------------------------------------------ */
-
-	async function setupWithPlot(): Promise<Ctx> {
-		const db = await openDatabase({ path: ":memory:" });
-		const repos = createRepos(db);
-		await repos.agents.upsert({
-			name: "refactor-bot",
-			renderedJson: { sections: { system: "x" } },
-		});
-		const project = await repos.projects.create({
-			gitUrl: "https://github.com/x/y.git",
-			localPath: "/data/projects/x/y",
-			defaultBranch: "main",
-			hasPlot: true,
-		});
-		const run = await repos.runs.create({
-			agentName: "refactor-bot",
-			projectId: project.id,
-			prompt: "p",
-			renderedAgentJson: {},
-			trigger: "manual",
-			burrowId: "bur_aaaaaaaaaaaa",
-			burrowRunId: "run_zzzzzzzzzzzz",
-		});
-		await repos.runs.markRunning(run.id);
-		return {
-			db,
-			repos,
-			broker: new RunEventBroker(),
-			runId: run.id,
-			projectPath: project.localPath,
-			workspacePath: "/data/burrow/ws",
-		};
-	}
-
-	test("authors a warren commit when project .plot/ has a delta the agent never committed (warren-343a)", async () => {
-		const plotCtx = await setupWithPlot();
-		try {
-			const f = fakeFs({
-				"/data/projects/x/y/.plot/plot-abc.events.jsonl":
-					'{"type":"run_dispatched","actor":"user:operator","at":"2026-05-18T10:00:00Z","data":{}}\n',
-				"/data/projects/x/y/.plot/plot-abc.json":
-					'{"id":"plot-abc","status":"active","updated_at":"2026-05-18T10:00:00Z"}',
-			});
-			const e = fakeExec({ stagedDelta: true });
-
-			const result = await reapRun({
-				runId: plotCtx.runId,
-				outcome: "succeeded",
-				repos: plotCtx.repos,
-				...reapDeps(fakeBurrowClient(makeBurrow()), { fs: f.fs, exec: e.exec }),
-				fs: f.fs,
-				exec: e.exec,
-			});
-
-			expect(result.plotCommitted).toBe(true);
-			expect(f.files.get("/data/burrow/ws/.plot/plot-abc.events.jsonl")).toContain(
-				"run_dispatched",
-			);
-			expect(f.files.get("/data/burrow/ws/.plot/plot-abc.json")).toContain('"status":"active"');
-			const gitArgs = e.calls.filter((c) => c.cmd === "git").map((c) => c.args);
-			// warren-c55e: per-carrier pathspecs (sorted), symmetric with seeds.
-			const paths = [".plot/plot-abc.events.jsonl", ".plot/plot-abc.json"];
-			expect(gitArgs).toContainEqual(["add", "--", ...paths]);
-			expect(gitArgs).toContainEqual(["diff", "--cached", "--quiet", "--", ...paths]);
-			const commit = gitArgs.find((a) => a[0] === "-c" && a.includes("commit"));
-			expect(commit).toEqual([
-				"-c",
-				"user.name=warren",
-				"-c",
-				"user.email=warren@os-eco.dev",
-				"commit",
-				"--no-verify",
-				"--only",
-				"-m",
-				"chore(warren): plot state",
-				"--",
-				...paths,
-			]);
-			const events = await plotCtx.repos.events.listByRun(plotCtx.runId);
-			expect(events.find((ev) => ev.kind === "reap.plot_committed")).toBeDefined();
-			// warren-035c: the commit spawn also pins the identity in env so an
-			// inherited GIT_AUTHOR_*/GIT_COMMITTER_* can't out-rank the `-c` config.
-			// warren-23dd: it also carries the repo-context GIT_* scrub (keys
-			// present-and-undefined) so a leaked GIT_DIR can't divert the commit.
-			const commitCall = e.calls.find((c) => c.cmd === "git" && c.args.includes("commit"));
-			expect(commitCall?.env).toEqual({
-				...gitRepoContextScrubEnv(),
-				...warrenCommitIdentityEnv(),
-			});
-			// warren-23dd: the non-commit git calls (add / diff --cached) in the
-			// same flow carry the scrub alone, mirroring clone-apply.ts.
-			const addCall = e.calls.find((c) => c.cmd === "git" && c.args[0] === "add");
-			expect(addCall?.env).toEqual(gitRepoContextScrubEnv());
-			const diffCall = e.calls.find(
-				(c) => c.cmd === "git" && c.args[0] === "diff" && c.args.includes("--cached"),
-			);
-			expect(diffCall?.env).toEqual(gitRepoContextScrubEnv());
-		} finally {
-			await plotCtx.db.close();
-		}
-	});
-
-	test("path-limits the plot commit to .plot/ so pre-staged unrelated files aren't swept in (warren-be12)", async () => {
-		const plotCtx = await setupWithPlot();
-		try {
-			const f = fakeFs({
-				"/data/projects/x/y/.plot/plot-abc.events.jsonl":
-					'{"type":"run_dispatched","actor":"user:operator","at":"2026-05-18T10:00:00Z","data":{}}\n',
-			});
-			const e = fakeExec({ stagedDelta: true });
-
-			await reapRun({
-				runId: plotCtx.runId,
-				outcome: "succeeded",
-				repos: plotCtx.repos,
-				...reapDeps(fakeBurrowClient(makeBurrow()), { fs: f.fs, exec: e.exec }),
-				fs: f.fs,
-				exec: e.exec,
-			});
-
-			const commit = e.calls
-				.filter((c) => c.cmd === "git")
-				.map((c) => c.args)
-				.find((a) => a[0] === "-c" && a.includes("commit"));
-			// `--only -- .plot/plot-abc.events.jsonl` confines the commit to the
-			// actually-copied carrier, so git ignores anything else an earlier
-			// step left in the index — even another file under `.plot/`
-			// (warren-c55e).
-			expect(commit).toContain("--only");
-			const dashDash = commit?.indexOf("--") ?? -1;
-			expect(dashDash).toBeGreaterThan(-1);
-			expect(commit?.slice(dashDash + 1)).toEqual([".plot/plot-abc.events.jsonl"]);
-		} finally {
-			await plotCtx.db.close();
-		}
-	});
-
-	test("does not commit when the agent already committed every .plot/ delta", async () => {
-		const plotCtx = await setupWithPlot();
-		try {
-			const f = fakeFs({
-				"/data/projects/x/y/.plot/plot-abc.events.jsonl":
-					'{"type":"run_dispatched","actor":"user:operator","at":"2026-05-18T10:00:00Z","data":{}}\n',
-			});
-			const e = fakeExec({ stagedDelta: false });
-
-			const result = await reapRun({
-				runId: plotCtx.runId,
-				outcome: "succeeded",
-				repos: plotCtx.repos,
-				...reapDeps(fakeBurrowClient(makeBurrow()), { fs: f.fs, exec: e.exec }),
-				fs: f.fs,
-				exec: e.exec,
-			});
-
-			expect(result.plotCommitted).toBe(false);
-			const commit = e.calls.find((c) => c.cmd === "git" && c.args.includes("commit"));
-			expect(commit).toBeUndefined();
-			const events = await plotCtx.repos.events.listByRun(plotCtx.runId);
-			expect(events.find((ev) => ev.kind === "reap.plot_committed")).toBeUndefined();
-		} finally {
-			await plotCtx.db.close();
-		}
-	});
-
-	test("skips .plot/.index.db* and non-plot-* entries when copying into the workspace", async () => {
-		const plotCtx = await setupWithPlot();
-		try {
-			const f = fakeFs({
-				"/data/projects/x/y/.plot/plot-abc.events.jsonl": '{"type":"note"}\n',
-				"/data/projects/x/y/.plot/.index.db": "binary-sqlite",
-				"/data/projects/x/y/.plot/.index.db-wal": "wal",
-				"/data/projects/x/y/.plot/README.md": "# docs",
-			});
-			const e = fakeExec({ stagedDelta: true });
-
-			await reapRun({
-				runId: plotCtx.runId,
-				outcome: "succeeded",
-				repos: plotCtx.repos,
-				...reapDeps(fakeBurrowClient(makeBurrow()), { fs: f.fs, exec: e.exec }),
-				fs: f.fs,
-				exec: e.exec,
-			});
-
-			expect(f.files.get("/data/burrow/ws/.plot/plot-abc.events.jsonl")).toBeDefined();
-			expect(f.files.get("/data/burrow/ws/.plot/.index.db")).toBeUndefined();
-			expect(f.files.get("/data/burrow/ws/.plot/.index.db-wal")).toBeUndefined();
-			expect(f.files.get("/data/burrow/ws/.plot/README.md")).toBeUndefined();
-		} finally {
-			await plotCtx.db.close();
-		}
-	});
-
-	test("trivial-merge child: warren commit keeps reap.empty_push silent", async () => {
-		const plotCtx = await setupWithPlot();
-		try {
-			const f = fakeFs({
-				"/data/projects/x/y/.plot/plot-abc.events.jsonl":
-					'{"type":"run_dispatched","actor":"user:operator","at":"2026-05-18T10:00:00Z","data":{}}\n',
-			});
-			const e = fakeExec({ stagedDelta: true, revListCount: "1" });
-
-			const result = await reapRun({
-				runId: plotCtx.runId,
-				outcome: "succeeded",
-				repos: plotCtx.repos,
-				...reapDeps(fakeBurrowClient(makeBurrow()), { fs: f.fs, exec: e.exec }),
-				fs: f.fs,
-				exec: e.exec,
-			});
-
-			expect(result.plotCommitted).toBe(true);
-			expect(result.branchPushed).toBe(true);
-			expect(result.commitsAhead).toBe(1);
-			const events = await plotCtx.repos.events.listByRun(plotCtx.runId);
-			expect(events.find((ev) => ev.kind === "reap.empty_push")).toBeUndefined();
-		} finally {
-			await plotCtx.db.close();
-		}
-	});
-
-	test("project without .plot/ skips the plot_commit step entirely", async () => {
-		const f = fakeFs({
-			"/data/projects/x/y/.plot/plot-abc.events.jsonl": '{"type":"x"}\n',
-		});
-		const e = fakeExec({ stagedDelta: true });
-
-		const result = await reapRun({
-			runId: ctx.runId,
-			outcome: "succeeded",
-			repos: ctx.repos,
-			...reapDeps(fakeBurrowClient(makeBurrow()), { fs: f.fs, exec: e.exec }),
-			fs: f.fs,
-			exec: e.exec,
-		});
-
-		expect(result.plotCommitted).toBe(false);
-		expect(f.files.get("/data/burrow/ws/.plot/plot-abc.events.jsonl")).toBeUndefined();
-		const gitArgs = e.calls.filter((c) => c.cmd === "git").map((c) => c.args);
-		expect(gitArgs.find((a) => a.includes("add"))).toBeUndefined();
-		expect(gitArgs.find((a) => a.includes("commit"))).toBeUndefined();
 	});
 
 	/* ------------------------------------------------------------------ */
