+++ src/runs/reap/test-helpers.ts
@@ -52,10 +52,6 @@ export function makeReapRunResult(overrides: Partial<ReapRunResult> = {}): ReapR
 		seedsClosed: 0,
 		seedsCreated: 0,
 		seedIdClosed: false,
-		plotEventsAppended: 0,
-		plotsUpdated: 0,
-		plotEventsMirrored: 0,
-		plotCommitted: false,
 		seedsCommitted: false,
 		branchPushed: false,
 		commitsAhead: null,
