+++ src/runs/reap/util.ts
@@ -26,10 +26,6 @@ export function buildAlreadyTerminalResult(run: RunRow): ReapRunResult {
 		seedsClosed: 0,
 		seedsCreated: 0,
 		seedIdClosed: false,
-		plotEventsAppended: 0,
-		plotsUpdated: 0,
-		plotEventsMirrored: 0,
-		plotCommitted: false,
 		seedsCommitted: false,
 		branchPushed: false,
 		commitsAhead: null,
