+++ src/runs/spawn/dispatch.ts
@@ -35,7 +35,6 @@
  * original error is rethrown for the caller (HTTP route, CLI).
  */
 
-import { join } from "node:path";
 import { NotFoundError, ValidationError } from "../../core/errors.ts";
 import { refreshProject } from "../../projects/manage.ts";
 import {
@@ -50,12 +49,6 @@ import { parseBurrowConfig } from "../burrow-config.ts";
 import { buildSeedFiles } from "../seed.ts";
 import { readCachedAgent, readProjectDefaults, resolveOverride } from "./agent-cache.ts";
 import { type EnvLike, injectWarrenCallbackEnv } from "./callback-env.ts";
-import {
-	defaultPlotAppender,
-	emitRunDispatchedToPlot,
-	extractModel,
-	resolveDispatcherHandle,
-} from "./plot-append.ts";
 import {
 	bindRunLogger,
 	logDispatched,
@@ -97,20 +90,6 @@ export async function spawnRun(input: SpawnRunInput): Promise<SpawnRunResult> {
 		});
 	}
 	const project = await input.repos.projects.require(input.projectId);
-	// warren-a8c3: gate plot_id on the project's hasPlot flag. Probed at
-	// addProject / refreshProjectClone time (warren-4e20). Refusing here
-	// keeps the runs row honest — a non-Plot project never grows a
-	// dangling plot_id that downstream PLOT_ID env injection (warren-e26f)
-	// or .plot/ mirroring (warren-7e0f) would have to second-guess.
-	if (input.plotId !== undefined && input.plotId !== "" && !project.hasPlot) {
-		throw new ValidationError(
-			`project ${project.id} has no .plot/ directory; plot_id is not accepted`,
-			{
-				recoveryHint:
-					"either omit plot_id on POST /runs, or run `plot init` in the project clone and refresh the project so warren picks up the .plot/ directory",
-			},
-		);
-	}
 	const baseAgent = readCachedAgent(agentRow.renderedJson, agentRow.name);
 	const burrowConfig = parseBurrowConfig(baseAgent.sections.burrow_config);
 
@@ -145,7 +124,7 @@ export async function spawnRun(input: SpawnRunInput): Promise<SpawnRunResult> {
 	const projectAfterRefresh = refreshed?.project ?? project;
 
 	// warren-c1a4: coordination project — host clone the post-dispatch
-	// seed stamp + Plot append target (defaults to the execution project).
+	// seed stamp target (defaults to the execution project).
 	const coordinationProject = await resolveCoordinationProject(
 		input.repos,
 		input.seedProjectId,
@@ -194,7 +173,6 @@ export async function spawnRun(input: SpawnRunInput): Promise<SpawnRunResult> {
 		renderedAgentJson: agent,
 		trigger: input.trigger ?? "manual",
 		...(input.seedId !== undefined ? { seedId: input.seedId } : {}),
-		...(input.plotId !== undefined && input.plotId !== "" ? { plotId: input.plotId } : {}),
 		...(input.mode !== undefined ? { mode: input.mode } : {}),
 		...(input.parentRunId !== undefined && input.parentRunId !== ""
 			? { parentRunId: input.parentRunId, cloneKind: input.cloneKind ?? "continue" }
@@ -218,21 +196,7 @@ export async function spawnRun(input: SpawnRunInput): Promise<SpawnRunResult> {
 		input.targetBranch,
 	);
 
-	// warren-e26f: when the run is bound to a Plot, inject the env vars the
-	// `plot` CLI inside the sandbox needs to identify itself. Gated on
-	// project.hasPlot (already validated above) AND a concrete plot_id on
-	// the run row — both must be set, otherwise we leave env empty so a
-	// non-Plot dispatch is byte-identical to the pre-change behavior. Actor
-	// shape is `agent:<agent-name>:<run-id>` per warren-000b SPEC §6 / Plot
-	// write-ACL contract. Run id is generated by runs.create above so it's
-	// already in hand.
-	const runEnv = composeRunEnv(
-		run.plotId,
-		agent.name,
-		run.id,
-		projectDefaults?.qualityGate,
-		input.serverEnv,
-	);
+	const runEnv = composeRunEnv(projectDefaults?.qualityGate, input.serverEnv);
 
 	// warren-b802: resolve per-project runtime override for the planner
 	// interactive agent at dispatch time so the agent row
@@ -311,23 +275,6 @@ export async function spawnRun(input: SpawnRunInput): Promise<SpawnRunResult> {
 				runId: run.id,
 				agentName: agent.name,
 				trigger: input.trigger,
-				now: input.now?.() ?? new Date(),
-			});
-		}
-		// warren-e848 / pl-2047 step 5: append a `run_dispatched` event to the
-		// originating Plot. Fire-and-log — same posture as writeSeedExtensions.
-		if (updated.plotId !== null && updated.plotId !== "") {
-			await emitRunDispatchedToPlot({
-				repos: input.repos,
-				runId: run.id,
-				plotDir: join(coordinationProject.localPath, ".plot"),
-				plotId: updated.plotId,
-				handle: resolveDispatcherHandle(input.dispatcherHandle),
-				agentName: agent.name,
-				model: extractModel(agent.frontmatter),
-				projectId: coordinationProject.id,
-				...(input.executionRepo !== undefined ? { executionRepo: input.executionRepo } : {}),
-				appender: input.plotAppender ?? defaultPlotAppender,
 				now: input.now?.() ?? new Date(),
 			});
 		}
@@ -411,20 +358,13 @@ async function resolveContinuationRef(
 
 // warren-b893: route Bun cache outside the workspace so `git add .` never sweeps it.
 const BUN_INSTALL_CACHE_DIR = "/tmp/bun-install-cache";
-/** Merge Plot env vars (warren-e26f), quality-gate (warren-5797), and Bun cache relocation
+/** Merge quality-gate (warren-5797) and Bun cache relocation
  * (warren-b893) into the sandbox env. Always returns a non-empty object. */
 function composeRunEnv(
-	plotId: string | null,
-	agentName: string,
-	runId: string,
 	qualityGate: string | undefined,
 	serverEnv?: Readonly<Record<string, string | undefined>>,
 ): Record<string, string> {
 	const env: Record<string, string> = { BUN_INSTALL_CACHE_DIR };
-	if (plotId !== null) {
-		env.PLOT_ID = plotId;
-		env.PLOT_ACTOR = `agent:${agentName}:${runId}`;
-	}
 	if (qualityGate !== undefined) env.WARREN_QUALITY_GATE = qualityGate;
 	// warren-f248: forward the warren API token + loopback URL so the agent
 	// can call back into warren's HTTP API (audit-warden delivery path).
