+++ src/runs/spawn/index.ts
@@ -4,7 +4,6 @@
  * pl-9088 step 6:
  *
  *   - `./dispatch.ts`         — `spawnRun` orchestrator + composers
- *   - `./plot-append.ts`      — `run_dispatched` Plot append (warren-e848)
  *   - `./seed-extensions.ts`  — post-dispatch seed extension write (pl-bb70)
  *   - `./agent-cache.ts`      — cached agent re-validation + override resolution
  *   - `./types.ts`            — `SpawnRunInput` / `SpawnRunResult` / appender shapes
@@ -14,14 +13,4 @@
  */
 
 export { composeDispatchPrompt, spawnRun } from "./dispatch.ts";
-export {
-	DEFAULT_DISPATCHER_HANDLE,
-	defaultPlotAppender,
-	resolveDispatcherHandle,
-} from "./plot-append.ts";
-export type {
-	AppendPlotRunDispatchedInput,
-	SpawnPlotAppender,
-	SpawnRunInput,
-	SpawnRunResult,
-} from "./types.ts";
+export type { SpawnRunInput, SpawnRunResult } from "./types.ts";
