+++ src/runs/watchdog.test-helpers.ts
@@ -30,10 +30,6 @@ export function fakeReapResult(state: ReapRunResult["state"]): ReapRunResult {
 		seedsClosed: 0,
 		seedsCreated: 0,
 		seedIdClosed: false,
-		plotEventsAppended: 0,
-		plotsUpdated: 0,
-		plotEventsMirrored: 0,
-		plotCommitted: false,
 		seedsCommitted: false,
 		branchPushed: false,
 		commitsAhead: null,
