+++ src/server/handlers/plots/intent.ts
@@ -8,12 +8,12 @@
 
 import { join } from "node:path";
 import { ValidationError } from "../../../core/errors.ts";
+import { resolveDispatcherHandle } from "../../../plots/dispatcher-handle.ts";
 import {
 	defaultPlotIntentEditor,
 	defaultPlotRenamer,
 	type EditPlotIntentPatch,
 } from "../../../plots/index.ts";
-import { resolveDispatcherHandle } from "../../../runs/index.ts";
 import { jsonResponse } from "../../response.ts";
 import type { RouteHandler, ServerDeps } from "../../types.ts";
 import { optionalString, readJsonBody, requireParam } from "../index.ts";
