+++ src/server/handlers/plots/workbench.ts
@@ -12,8 +12,8 @@
 import { join } from "node:path";
 import { NotFoundError, ValidationError } from "../../../core/errors.ts";
 import { ProjectLacksPlotError } from "../../../plan-runs/errors.ts";
+import { resolveDispatcherHandle } from "../../../plots/dispatcher-handle.ts";
 import { defaultPlotQuestionAnswerer } from "../../../plots/index.ts";
-import { resolveDispatcherHandle } from "../../../runs/index.ts";
 import { jsonResponse } from "../../response.ts";
 import type { RouteHandler, ServerDeps } from "../../types.ts";
 import { optionalString, readJsonBody, requireParam } from "../index.ts";
